backend/requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3
flask-wtf==1.2.1
email-validator==2.1.0

# ASGI Server
uvicorn==0.27.0
asgiref==3.10.0

# Machine Learning
onnxruntime>=1.16.0
Pillow>=10.4.0
numpy>=1.24.0

# Web & API
requests==2.31.0
openai==1.12.0
gunicorn==23.0.0

# Geo & Location
geocoder==1.38.1
geoip2==4.7.0
python-ipware==3.0.0
geopy==2.4.1

# Utilities
python-dotenv==1.0.0
user-agents==2.2.0

# Export & Reporting
reportlab==4.4.7